Change original git project to "noabody" fork
As the original mmozeiko project has not been updated for 2 years, some functions such as the interpretation of the verbose output are obsolete. In noabody's fork that and other problems have been solved, also has other new features such as the option to see the "Shared" area of your Drive account with a simple checkbox.